+++ src/PodmanClient/PodmanClient.Generate.cs
@@ -72,7 +72,7 @@ public partial class PodmanClient {
         ("start", start.ToString().ToLowerInvariant()),
         ("logDriver", logDriver),
       ],
-      cancellationToken
+      cancellationToken: cancellationToken
     );
   }
 }