+++ src/PodmanClient/PodmanClient.Images.Api.cs
@@ -152,7 +152,7 @@ public partial class PodmanClient {
         ("reference", reference),
         ("url", url),
       ],
-      cancellationToken
+      cancellationToken: cancellationToken
     );
   }